+++ examples/intro_sync.cpp
@@ -7,7 +7,7 @@
 #include <tuple>
 #include <string>
 #include <thread>
-#include <boost/asio/io_context.hpp>
+#include <boost/asio.hpp>
 #include <aedis.hpp>
 
 // Include this in no more than one .cpp file.
@@ -17,34 +17,44 @@ namespace net = boost::asio;
 using aedis::adapt;
 using aedis::resp3::request;
 using aedis::endpoint;
-using connection = aedis::sync<aedis::connection<>>;
+using connection = aedis::connection<>;
+
+template <class Adapter>
+auto exec(connection& conn, request const& req, Adapter adapter, boost::system::error_code& ec)
+{
+   net::dispatch(
+      conn.get_executor(),
+      net::deferred([&]() { return conn.async_exec(req, adapter, net::deferred); }))
+      (net::redirect_error(net::use_future, ec)).get();
+}
+
+auto logger = [](auto const& ec)
+   { std::clog << "Run: " << ec.message() << std::endl; };
 
 int main()
 {
    try {
       net::io_context ioc{1};
-      auto work = net::make_work_guard(ioc);
-      std::thread t1{[&]() { ioc.run(); }};
 
-      connection conn{work.get_executor()};
-      std::thread t2{[&]() {
-         boost::system::error_code ec;
-         endpoint ep{"127.0.0.1", "6379"};
-         conn.run(ep, ec);
+      connection conn{ioc};
+      std::thread t{[&]() {
+         conn.async_run({"127.0.0.1", "6379"}, logger);
+         ioc.run();
       }};
 
       request req;
       req.push("PING");
       req.push("QUIT");
 
+      boost::system::error_code ec;
       std::tuple<std::string, aedis::ignore> resp;
-      conn.exec(req, adapt(resp));
-      std::cout << "Response: " << std::get<0>(resp) << std::endl;
+      exec(conn, req, adapt(resp), ec);
 
-      work.reset();
+      std::cout
+         << "Exec: " << ec.message() << "\n"
+         << "Response: " << std::get<0>(resp) << std::endl;
 
-      t1.join();
-      t2.join();
+      t.join();
    } catch (std::exception const& e) {
       std::cerr << e.what() << std::endl;
    }